[libra framework][move-prover][DRAFT] Refactor Roles.move and clients
Current status: This is still a draft. I've now removed all uses of
LibraRootRole.  It passes the prover's tests, but note the move-lang
functional tests. For THOSE, the first one is crashing on an "unwrap"
error: translated_ir_tests/block/expried_transaction.move The file
doesn't do anything, so it's hard to see how this is related to
something I did in Move files. Tim: Can you help me understand this?

This is a partial refactor of Roles.move and modules that depend on it
to make it easier to verify access control policies. I've tried to
pick out representative examples of how to do it. Completing the
refactor should be straightforward matter of repeating the same
patterns (I'm hesitating to do that just because I'm going to need to
touch a lot of files and want to be careful about merge conflicts).

The previous approach of publishing Privilege<Priv> objects and
extacting capabilities from them created several problems for the Move
Prover. The most fundamental of these is that the Move Prover cannot
prove a property unless it sees all of the code that could invalidate
that property. As currently designed, that means the module containing
the property and all its dependencies.  We feel that this is highly
desireable, as compared with assuming we can formally verify all
modules, because it is more compatible with the eventual goal of
opening blockchain to others who want to develop Move modules that we
may not completely trust.  In the presence of untrusted modules,
properties that have been proven valid using modular verification will
still be guaranteed.

More pragmatically, there were a number of other problems, ranging
from tricky verification problems that we haven't yet solved to
limitations of the verifier that we don't have time to fix.  The
refactor should eliminate the need to prove many properties and will
make others much simpler to prove.

Currently, a Move function that is restricted to certain signers
has an argument of type &Capability<Priv>, which can only be generated
if there is a signer of the transaction with a published Privilege<Priv>.

The rewrite replaces that capability with an account: &Signer
argument, and calls assert(has_..._privilege(account), <abort code>)
to abort the function if the signer does not have the required
privilege.  As it happens, has_..._privilege is always a property of
the role of the address as defined by the RoleId in Roles.move, so we
don't actually need to have or publish Privilege<Priv> objects.  To
achieve obvious, visible correspondence to our spreadsheet of
"governance roles and privileges", these functions have names
corresponding to privileges from the spreadsheet.  Several identical
functions have the same body, because some roles have multiple
privileges. (In the future, perhaps we can auto-generate verification
properties from that spreadsheet.)

This approach retains the flexibility of changing the addresses that
have a given role, because the RoleId implementation is unchanged.
The correspondence between roles and privileges can also be changed
easily by changing the appropriate has_..._privilege functions. Right
now, they all just check for a particular role ID, but there would be
no problem changing them to allow multiple alternative roles, or to
check for the presence of a published resource of some type, or lots
of other policies.

It looks like every capability in Roles can be replaced in this way.
There are other capabilities in the system that are implemented
outside of Roles (mint/burn/preburn and config, for example). It is
possible that they will also present hard verification problems and
that we should consider refactoring them.  At this point, I want to
finish the Roles refactor before thinking about that problem very
hard.