[DON'T SUBMIT AS PR]
This contains a couple of small changes to Lera's PR 6403.

1. Add error category to her new assert in
update_config_and_reconfigure.  I guessed that "LIMIT_EXCEEDED" was
the best one, but maybe "INVALID_STATE" would be better.  Whichever we
go with, the documentation in set_validator_config_and_reconfigure
needs another table row to discuss this new abort.

2. Add new aborts_if for the new assert.  Which, unfortunately, still
times out (except that there is a pragma verify_duration_estimate in
the spec, which turns off verification for
update_config_and_reconfigure) In LibraSystem, correct asserts usually
finish quickly, and incorrect ones often time out (more time probably
won't help -- they're stuck). The aborts_if is tricky, so I'm not sure
whether the problem is a mistake or just that the problem is too hard
for the Prover.

3. Added a similar aborts_if to set_validator_config_and_reconfigure.
I'm not sure whether I can just "include" a schema from
LibraSystem.move, because the is_validator_updated flag is defined
differently in set_validator_config_and_reconfigure than in
LibraSystem.move.  I recently added assertions to
set_validator_config_and_reconfigure, and this *particular* assertion
took me a full day to debug, so I'm playing it safe.  This aborts_if
verifies quickly because it's similar to the aborts_if in
LibraSystem::update_config_and_reconfigure.  This success does not
mean that the aborts_if in set_validator_config_and_reconfigure is
correct, just that it's consistent with the aborts_if in LibraSystem.